feat: Use Struct
